+++ cloudsmith_cli/cli/commands/__init__.py
@@ -3,11 +3,14 @@ from __future__ import absolute_import, print_function, unicode_literals
 
 from .main import main  # noqa
 from . import check  # noqa
+from . import copy  # noqa
 from . import delete  # noqa
 from . import docs  # noqa
 from . import help_  # noqa
 from . import list_  # noqa
+from . import move  # noqa
 from . import push  # noqa
+from . import resync  # noqa
 from . import status  # noqa
 from . import token  # noqa
 from . import whoami  # noqa